discard
Discord is a free, ad-free, decentralized social platform that allows users to communicate through voice, video, and text chat, as well as share media and files. Users can connect with others through direct chat or servers, which can be public or private. Discord is popular with gamers, but it's also used for other purposes, such as building communities, hosting events, and rewarding customers